fix(security): resolve 4 CodeQL alerts
- Update gradle-build-action from vulnerable SHA to v2.4.2
- Add minimal contents:read permissions to release.yml jobs
- Widen RadarLeafSilhouetteIndex.vertexBase to int to avoid
  implicit short narrowing in compound assignment

+++ infinity/src/main/java/infinity/client/view/RadarLeafSilhouetteIndex.java
@@ -99,7 +99,7 @@ public final class RadarLeafSilhouetteIndex {
         final FloatBuffer positions = BufferUtils.createFloatBuffer(columnCount * 4 * 3);
         final ShortBuffer indices = BufferUtils.createShortBuffer(columnCount * 6);
 
-        short vertexBase = 0;
+        int vertexBase = 0;
         for (int cx = 0; cx < sizeX; cx++) {
             for (int cz = 0; cz < sizeZ; cz++) {
                 if (!anySolidInColumn(cells, cx, cz, sizeY)) {
